requirements.txt
Flask
gunicorn
/Users/bartes/Projects/docker-development/repos/castle-python/dist/castle-6.1.0.tar.gz
python-dotenv
